# project/api/openapi.yaml
openapi: 3.1.0

info:
  title: Users Admin API
  version: 0.1.0
  summary: Admin API for managing users, their phones, and their addresses.
  license:
    name: MIT
    identifier: MIT
  description: |
    A portfolio project built **design-first**: this contract is written and reviewed
    before any server code exists.

    Status: all 6 operations work, stored in PostgreSQL.

    **Demo only: all data is fake and resets every night at 08:00 UTC.**

    **Trying it on the hosted demo:** call `POST /demo/token` below, copy the `token`, click **Authorize**, and paste it. Calls without it get `401`, which is auth doing its job.

    ## Conventions
    - **Auth:** every operation requires an admin token (`Authorization: Bearer <JWT>`) with an expiry (`exp`). The only exceptions are this docs page, the spec, and the demo token. Auth is checked first: without a valid token, every request gets `401` (or `403` for a non-admin), whatever else is wrong with it.
    - **Body size:** request bodies over 100 KB return `413`.
    - **Errors:** all errors use Problem Details (RFC 9457), `application/problem+json`.
    - **Whole user:** phones and addresses are part of the user. Create and update send all of them; there are no separate phone or address endpoints.
    - **Versioning:** responses carry an `ETag`; updates and deletes must send it back in `If-Match`.
    - **Paging:** list endpoints take `page` and `pageSize` (max 100) and return `{ items, page, pageSize, totalItems, totalPages }`.
    - **Dates:** timestamps are UTC ISO 8601; date of birth is a date only (`YYYY-MM-DD`).
    - **Rate limits:** 100 requests per minute per client IP address, counted in fixed one-minute windows, including requests without a valid token. Exceeding it returns `429` with a `Retry-After` header (seconds until the next window).
    - **Unknown input:** unknown body fields and unknown query parameters return `400`, naming each one.
    - **Unknown paths and methods:** an unknown path returns `404`; a known path with an unsupported method returns `405` with an `Allow` header.

tags:
  - name: Demo
    description: Only on the hosted demo. Get a 1-hour admin token to try the API.
  - name: Users
    description: Create, list, update, delete, and restore users, including their phones and addresses.

servers:
  - url: /
    description: Wherever this page is served from (your laptop or Render)

security:
  - adminToken: []

paths:
  /demo/token:
    post:
      tags: [Demo]
      operationId: getDemoToken
      summary: Get a demo admin token
      description: |
        **Only when the server runs in demo mode** (the hosted demo). Anywhere else this path doesn't exist, so without a token it answers `401` like any other path.
        No token or body needed. The token is an admin token that expires after 1 hour; paste it into **Authorize**.
        Demo only: all data is fake and resets every night at 08:00 UTC.
      security: []
      responses:
        "200":
          description: "A new admin token. Never cached (`Cache-Control: no-store`)."
          headers:
            Cache-Control:
              description: Always `no-store`, so no browser or proxy keeps a copy of the token.
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoToken"
        "405":
          description: A method other than `POST`. The `Allow` header lists `POST`.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /v1/users:
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      description: |
        Email must not belong to any user, **including deleted ones** (emails are never reused).
        Date of birth must not be in the future, and the user must be 18 or older.
        Phones and addresses: 1–3 of each, one per type, exactly one primary (see `UserInput`).
        If the server has a user limit and it's reached (deleted users count), the create is refused with `409` and type `/problems/user-limit`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: The email is already used, or the server's user limit is reached.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                emailTaken:
                  summary: Email already used
                  value:
                    type: /problems/conflict
                    title: Conflict
                    status: 409
                    detail: A deleted user has this email; restore that user instead
                userLimit:
                  summary: User limit reached
                  value:
                    type: /problems/user-limit
                    title: User limit reached
                    status: 409
                    detail: This server stores at most 200 users, including deleted ones
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: |
        Returns one page of users in the basic view.
        `search` is a partial, case-insensitive match against first name, last name, and email.
        Sorting ignores case; ties are broken by first name, then id, always ascending (`order` applies only to the `sort` field). Deleted users are hidden unless `includeDeleted=true`.
        A page past the end returns `200` with empty `items` and the real `totalItems` and `totalPages`.
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/IncludeDeleted"
      responses:
        "200":
          description: One page of users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /v1/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"

    get:
      tags: [Users]
      operationId: getUser
      summary: Get one user
      description: Basic view by default; `view=detailed` adds date of birth, phones, addresses, version, and timestamps.
      parameters:
        - $ref: "#/components/parameters/View"
        - $ref: "#/components/parameters/IncludeDeleted"
      responses:
        "200":
          description: The user.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UserBasic"
                  - $ref: "#/components/schemas/UserDetailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      tags: [Users]
      operationId: replaceUser
      summary: Update a user (replace)
      description: |
        Replaces the whole user, including phones and addresses. A phone or address left out is removed permanently.
        Same rules as create. Send the `ETag` you loaded in `If-Match`.
        A deleted user must be restored before it can be updated.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Updated.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete a user
      description: |
        Marks the user as deleted; nothing is removed. Their phones and addresses become hidden.
        Send the `ETag` you loaded in `If-Match`.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /v1/users/{userId}/restore:
    parameters:
      - $ref: "#/components/parameters/UserId"

    post:
      tags: [Users]
      operationId: restoreUser
      summary: Restore a deleted user
      description: Brings back the user along with their phones and addresses. Restoring a user who isn't deleted returns `409`.
      responses:
        "200":
          description: Restored.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"

components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A JWT whose `role` claim is `admin`.

  schemas:
    DemoToken:
      type: object
      additionalProperties: false
      required: [token, tokenType, expiresIn, note]
      properties:
        token:
          type: string
          description: A JWT with the `admin` role.
          example: eyJhbGciOiJIUzI1NiJ9.eyJyb2xlIjoiYWRtaW4ifQ.signature
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          description: Seconds until the token stops working.
          const: 3600
        note:
          type: string
          example: "Demo only: all data is fake and resets every night at 08:00 UTC"

    Uuid:
      type: string
      format: uuid
      example: 3f2a8c1e-7b4d-4e2a-9c6f-1d5e8b7a0c42

    Timestamp:
      type: string
      format: date-time
      description: UTC, ISO 8601.
      example: "2026-09-12T18:40:00Z"

    Version:
      type: integer
      minimum: 1
      description: Increments on every change. Also sent as the `ETag` header.
      example: 3

    PersonName:
      type: string
      minLength: 1
      maxLength: 50
      pattern: "^(?=.*[A-Za-z])[A-Za-z .'-]+$"
      description: |
        1–50 characters after trimming surrounding spaces. Plain English letters (A–Z, a–z),
        spaces, hyphens, periods, and the straight apostrophe ('). Must contain at least one letter.
      example: O'Brien

    Email:
      type: string
      format: email
      maxLength: 254
      description: Unique across all users, including deleted ones. Uniqueness ignores case.
      example: jose.garcia@example.com

    DateOfBirth:
      type: string
      format: date
      description: Date only (`YYYY-MM-DD`). Not in the future; the user must be 18 or older.
      example: "1990-05-17"

    UserInput:
      type: object
      additionalProperties: false
      description: Unknown fields (such as `id` or `role`) are rejected with `400`.
      required: [firstName, lastName, email, dateOfBirth, phones, addresses]
      properties:
        firstName:
          $ref: "#/components/schemas/PersonName"
        lastName:
          $ref: "#/components/schemas/PersonName"
        email:
          $ref: "#/components/schemas/Email"
        dateOfBirth:
          $ref: "#/components/schemas/DateOfBirth"
        phones:
          type: array
          minItems: 1
          maxItems: 3
          description: |
            One per type. A single phone is primary automatically.
            With 2–3, exactly one must have `primary: true`. Breaking either rule returns `400`.
          items:
            $ref: "#/components/schemas/PhoneInput"
        addresses:
          type: array
          minItems: 1
          maxItems: 3
          description: |
            One per type. A single address is primary automatically.
            With 2–3, exactly one must have `primary: true`. Breaking either rule returns `400`.
          items:
            $ref: "#/components/schemas/AddressInput"
      example:
        firstName: Jose
        lastName: Garcia-Lopez
        email: jose.garcia@example.com
        dateOfBirth: "1990-05-17"
        phones:
          - number: (305) 555-1234
            type: mobile
        addresses:
          - street: 123 Main St
            street2: Apt 4B
            city: Miami
            state: FL
            zip: "33101"
            type: home

    UserBasic:
      type: object
      required: [id, firstName, lastName, email]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        firstName:
          $ref: "#/components/schemas/PersonName"
        lastName:
          $ref: "#/components/schemas/PersonName"
        email:
          $ref: "#/components/schemas/Email"
        deletedAt:
          description: When the user was deleted; `null` if active. Only included when `includeDeleted=true`.
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
      example:
        id: 3f2a8c1e-7b4d-4e2a-9c6f-1d5e8b7a0c42
        firstName: Jose
        lastName: Garcia-Lopez
        email: jose.garcia@example.com

    UserDetailed:
      allOf:
        - $ref: "#/components/schemas/UserBasic"
        - type: object
          required: [dateOfBirth, phones, addresses, version, createdAt, updatedAt]
          properties:
            dateOfBirth:
              $ref: "#/components/schemas/DateOfBirth"
            phones:
              type: array
              minItems: 1
              maxItems: 3
              description: Primary first, then by type in this order - mobile, home, work.
              items:
                $ref: "#/components/schemas/Phone"
            addresses:
              type: array
              minItems: 1
              maxItems: 3
              description: Primary first, then by type in this order - home, work, mailing.
              items:
                $ref: "#/components/schemas/Address"
            version:
              $ref: "#/components/schemas/Version"
            createdAt:
              $ref: "#/components/schemas/Timestamp"
            updatedAt:
              $ref: "#/components/schemas/Timestamp"
      example:
        id: 3f2a8c1e-7b4d-4e2a-9c6f-1d5e8b7a0c42
        firstName: Jose
        lastName: Garcia-Lopez
        email: jose.garcia@example.com
        dateOfBirth: "1990-05-17"
        phones:
          - number: "+13055551234"
            type: mobile
            primary: true
        addresses:
          - street: 123 Main St
            street2: Apt 4B
            city: Miami
            state: FL
            zip: "33101"
            type: home
            primary: true
        version: 1
        createdAt: "2026-09-12T18:40:00Z"
        updatedAt: "2026-09-12T18:40:00Z"

    PhoneType:
      type: string
      enum: [mobile, home, work]

    PhoneInput:
      type: object
      additionalProperties: false
      required: [number, type]
      properties:
        number:
          type: string
          minLength: 10
          maxLength: 20
          description: |
            A U.S. number in any common format: `3055551234`, `(305) 555-1234`, `305.555.1234`, `+1 305 555 1234`.
            Must be 10 digits (after an optional leading `+1` or `1`); the area code can't start with 0 or 1.
            Stored and returned as E.164 (`+13055551234`).
          example: (305) 555-1234
        type:
          $ref: "#/components/schemas/PhoneType"
        primary:
          type: boolean
          description: Optional when it's the only phone (primary automatically). With 2–3 phones, exactly one must be `true`.
      example:
        number: (305) 555-1234
        type: mobile
        primary: true

    Phone:
      type: object
      required: [number, type, primary]
      properties:
        number:
          type: string
          pattern: "^\\+1[2-9]\\d{9}$"
          description: E.164 format.
          example: "+13055551234"
        type:
          $ref: "#/components/schemas/PhoneType"
        primary:
          type: boolean
          example: true

    AddressType:
      type: string
      enum: [home, work, mailing]

    UsState:
      type: string
      description: The 50 states, DC, and the territories AS, GU, MP, PR, and VI.
      enum: [AL, AK, AZ, AR, CA, CO, CT, DE, FL, GA, HI, ID, IL, IN, IA, KS, KY, LA, ME, MD,
             MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC,
             SD, TN, TX, UT, VT, VA, WA, WV, WI, WY, DC, AS, GU, MP, PR, VI]
      example: FL

    AddressText100:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^[A-Za-z0-9 ,.#-]+$"
      description: Letters, digits, spaces, and `- , . #`.
      example: 123 Main St.

    AddressInput:
      type: object
      additionalProperties: false
      required: [street, city, state, zip, type]
      properties:
        street:
          $ref: "#/components/schemas/AddressText100"
        street2:
          allOf:
            - $ref: "#/components/schemas/AddressText100"
          description: Optional (apartment, suite, unit). Omit for none.
          example: Apt 4B
        city:
          type: string
          minLength: 1
          maxLength: 50
          pattern: "^[A-Za-z0-9 ,.#-]+$"
          description: Letters, digits, spaces, and `- , . #`.
          example: Miami
        state:
          $ref: "#/components/schemas/UsState"
        zip:
          type: string
          pattern: "^\\d{5}$"
          description: Exactly 5 digits; leading zeros kept (e.g. `02134`).
          example: "33101"
        type:
          $ref: "#/components/schemas/AddressType"
        primary:
          type: boolean
          description: Optional when it's the only address (primary automatically). With 2–3 addresses, exactly one must be `true`.
      example:
        street: 123 Main St
        street2: Apt 4B
        city: Miami
        state: FL
        zip: "33101"
        type: home
        primary: true

    Address:
      type: object
      required: [street, street2, city, state, zip, type, primary]
      properties:
        street:
          type: string
          example: 123 Main St
        street2:
          type: [string, "null"]
          description: "`null` if none."
          example: Apt 4B
        city:
          type: string
          example: Miami
        state:
          $ref: "#/components/schemas/UsState"
        zip:
          type: string
          example: "33101"
        type:
          $ref: "#/components/schemas/AddressType"
        primary:
          type: boolean
          example: true

    UserPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/UserBasic"

    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        totalItems:
          type: integer
          minimum: 0
          example: 57
        totalPages:
          type: integer
          minimum: 0
          example: 3

    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: dateOfBirth
        message:
          type: string
          example: Must be 18 or older

    Problem:
      type: object
      description: Problem Details (RFC 9457).
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: Identifies the kind of problem.
          example: /problems/validation
        title:
          type: string
          description: Short, human-readable summary of the problem kind.
          example: Invalid input
        status:
          type: integer
          description: The HTTP status code.
          example: 400
        detail:
          type: string
          description: Explanation specific to this occurrence. Never contains stack traces, SQL, or file paths.
          example: 2 fields are invalid
        instance:
          type: string
          format: uri-reference
          description: The request path this problem occurred on.
          example: /v1/users
        errors:
          type: array
          description: Field-level problems (validation failures only).
          items:
            $ref: "#/components/schemas/FieldError"

  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"

    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1

    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

    Search:
      name: search
      in: query
      required: false
      description: Partial, case-insensitive match against first name, last name, and email. `%` and `_` are plain characters.
      schema:
        type: string
        minLength: 1
        maxLength: 100
        example: garc

    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [lastName, email]
        default: lastName

    Order:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: asc

    View:
      name: view
      in: query
      required: false
      description: "`detailed` adds date of birth, phones, addresses, version, and timestamps."
      schema:
        type: string
        enum: [basic, detailed]
        default: basic

    IncludeDeleted:
      name: includeDeleted
      in: query
      required: false
      description: Include users marked as deleted.
      schema:
        type: boolean
        default: false

    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The `ETag` from when the record was loaded. A stale value returns `412`.
      schema:
        type: string
        example: '"3"'

  headers:
    ETag:
      description: The record's current version, quoted.
      schema:
        type: string
        example: '"3"'

    Location:
      description: Path of the newly created record.
      schema:
        type: string
        format: uri-reference
        example: /v1/users/3f2a8c1e-7b4d-4e2a-9c6f-1d5e8b7a0c42

    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        example: 30

    WWWAuthenticate:
      description: How to authenticate.
      schema:
        type: string
        example: Bearer

  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/validation
            title: Invalid input
            status: 400
            detail: 2 fields are invalid
            instance: /v1/users
            errors:
              - field: email
                message: Must be a valid email
              - field: dateOfBirth
                message: Must be 18 or older

    Unauthorized:
      description: Missing, invalid, or expired token.
      headers:
        WWW-Authenticate:
          $ref: "#/components/headers/WWWAuthenticate"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/unauthorized
            title: Not signed in
            status: 401
            detail: A valid admin token is required

    Forbidden:
      description: Valid token, but not an admin.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/forbidden
            title: Not allowed
            status: 403
            detail: This operation requires the admin role

    NotFound:
      description: No such record, or it is hidden because the user is deleted.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/not-found
            title: Not found
            status: 404
            detail: No user with this id

    Conflict:
      description: Breaks a uniqueness rule, or the action doesn't fit the record's state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/conflict
            title: Conflict
            status: 409
            detail: A deleted user has this email; restore that user instead

    PreconditionFailed:
      description: The `If-Match` version is out of date.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/version-mismatch
            title: Version out of date
            status: 412
            detail: This record changed since it was loaded; reload and try again

    PayloadTooLarge:
      description: The request body is over 100 KB. It is refused without being read.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/too-large
            title: Body too large
            status: 413
            detail: Request body must be 100 KB or smaller

    PreconditionRequired:
      description: The `If-Match` header is missing.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/version-required
            title: Version required
            status: 428
            detail: Send the If-Match header with the version you loaded

    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: /problems/rate-limited
            title: Too many requests
            status: 429
            detail: Rate limit exceeded; retry after 30 seconds